COMMIT MESSAGE I am committing several files.
blog-post is where the full text is for the post
blog-template was the page I made that serves as a TOC for all the blog posts.  I need to do some more work here.
stylesheet is the css for blog-template.
post-stylesheet is the css for blog-post.
the other files are the blogs md and html files.